Refactoring of organiztion model, statistics controller
