+++ app/controllers/statistics_controller.rb
@@ -1,70 +1,57 @@
 class StatisticsController < ApplicationController
   layout 'settings'
+  respond_to :json, except: :index
 
   def index
   end
 
   def balance
-    if params[:scale].present? && params[:step].present?
-      data = current_organization.data_balance(params[:scale], params[:step])
-    elsif params[:scale].present?
-      data = current_organization.data_balance(params[:scale])
-    elsif params[:step].present?
-      data = current_organization.data_balance('months', params[:step])
-    else
-      data = current_organization.data_balance
-    end
-    respond_to do |format|
-      format.json { render json: data }
-    end
+    params[:scale] = 'months' if params[:scale].blank?
+    params[:step] = 0 if params[:step].blank?
+    data = StatisticData::ColumnsChart.new(current_organization)
+      .data_balance(params[:scale], params[:step])
+    respond_with(data)
   end
 
   def customers_chart
-    data = current_organization.customers_by_months(params[:type])
-    respond_to do |format|
-      format.json { render json: data }
-    end
+    data = StatisticData::ColumnsChart.new(current_organization)
+      .customers_by_months(params[:type])
+    respond_with(data)
   end
 
   def income_by_customers
-    incomes = current_organization.by_customers(:incomes, params[:period])
-    respond_to do |format|
-      format.json { render json: incomes }
-    end
+    incomes = StatisticData::RoundChart.new(current_organization)
+      .by_customers(:incomes, params[:period])
+    respond_with(incomes)
   end
 
   def expense_by_customers
-    expenses = current_organization.by_customers(:expenses, params[:period])
-    respond_to do |format|
-      format.json { render json: expenses }
-    end
+    expenses = StatisticData::RoundChart.new(current_organization)
+      .by_customers(:expenses, params[:period])
+    respond_with(expenses)
   end
 
   def totals_by_customers
-    totals = current_organization.totals_by_customers(params[:period])
-    respond_to do |format|
-      format.json { render json: totals }
-    end
+    totals = StatisticData::RoundChart.new(current_organization)
+      .totals_by_customers(params[:period])
+    respond_with(totals)
   end
 
   def balances_by_customers
-    balances = current_organization.balances_by_customers(params[:period])
-    respond_to do |format|
-      format.json { render json: balances }
-    end
+    balances = StatisticData::ColumnsChart.new(current_organization)
+      .balances_by_customers(params[:period])
+    respond_with(balances)
   end
 
   def income_by_categories
-    incomes = current_organization.by_categories(:incomes, params[:period])
-    respond_to do |format|
-      format.json { render json: incomes }
-    end
+    incomes = StatisticData::RoundChart.new(current_organization)
+      .by_categories(:incomes, params[:period])
+    respond_with(incomes)
   end
 
   def expense_by_categories
-    expenses = current_organization.by_categories(:expenses, params[:period])
-    respond_to do |format|
-      format.json { render json: expenses }
-    end
+    expenses = StatisticData::RoundChart.new(current_organization)
+      .by_categories(:expenses, params[:period])
+    respond_with(expenses)
   end
 end
