+++ app/models/invoice_item.rb
@@ -16,6 +16,7 @@
 
 class InvoiceItem < ApplicationRecord
   include CustomerConcern
+  include Period
   customer_concern_callbacks
 
   belongs_to :invoice, inverse_of: :invoice_items
@@ -30,25 +31,4 @@ class InvoiceItem < ApplicationRecord
   validates :amount, numericality: { greater_than: 0,
     less_than_or_equal_to: Dictionaries.money_max }
   validates :hours, numericality: { greater_than: 0 }, allow_nil: true
-
-  private
-
-  def self.period(period)
-    case period
-    when 'current-month'
-      where('DATE(invoice_items.date) between ? AND ?', Date.current.beginning_of_month, Date.current.end_of_month)
-    when 'last-3-months'
-      where('DATE(invoice_items.date) between ? AND ?', (Date.current - 3.months).beginning_of_day, Date.current.end_of_month)
-    when 'prev-month'
-      prev_month_begins = Date.current.beginning_of_month - 1.months
-      where('DATE(invoice_items.date) between ? AND ?', prev_month_begins,
-        prev_month_begins.end_of_month)
-    when 'this-year'
-      where('DATE(invoice_items.date) between ? AND ?', Date.current.beginning_of_year, Date.current.end_of_year)
-    when 'quarter'
-      where('DATE(invoice_items.date) between ? AND ?', Date.current.beginning_of_quarter, Date.current.end_of_quarter)
-    else
-      all
-    end
-  end
 end
