+++ app/models/transaction.rb
@@ -23,6 +23,7 @@ require "./lib/time_range.rb"
 class Transaction < ApplicationRecord
   include MoneyRails::ActionViewExtension
   include TimeRange
+  include Period
   TRANSACTION_TYPES = %w(Residue)
 
   acts_as_paranoid
@@ -226,25 +227,6 @@ class Transaction < ApplicationRecord
 
   def update_invoice_paid_at
     self.organization.invoices.where(id: self.invoice_id).first.try(:update, {paid_at: self.date})
-  end
-
-  def self.period(period)
-    case period
-    when 'current-month'
-      where('DATE(transactions.date) between ? AND ?', Date.current.beginning_of_month, Date.current.end_of_month)
-    when 'last-3-months'
-      where('DATE(transactions.date) between ? AND ?', (Date.current - 3.months).beginning_of_day, Date.current.end_of_month)
-    when 'prev-month'
-      prev_month_begins = Date.current.beginning_of_month - 1.months
-      where('DATE(transactions.date) between ? AND ?', prev_month_begins,
-        prev_month_begins.end_of_month)
-    when 'this-year'
-      where('DATE(transactions.date) between ? AND ?', Date.current.beginning_of_year, Date.current.end_of_year)
-    when 'quarter'
-      where('DATE(transactions.date) between ? AND ?', Date.current.beginning_of_quarter, Date.current.end_of_quarter)
-    else
-      all
-    end
   end
 
   def self.date_from(from)
